Fix method names in Broadcastable "same as" docs
Many of the methods in the Broadcastable concern are convenience
methods e.g. same as one of the other methods but with the stream
automatically set to the current model, or same as one of the
other methods but run asynchronously.

It looked like the equivalent method names weren't right in some
of the documentation comments. This commit corrects these.